Backend: migrate stats models to seperate app. Update everything
