+++ backend/api/constants.py
@@ -53,49 +53,12 @@ QUESTION_ANSWER_CHOICE_LIST = [
     "abcd",
 ]
 
-FEEDBACK_LIKE = "like"
-FEEDBACK_DISLIKE = "dislike"
-FEEDBACK_CHOICES = [
-    (FEEDBACK_LIKE, "Positif"),
-    (FEEDBACK_DISLIKE, "Négatif"),
-]
-
-QUESTION_SOURCE_QUESTION = "question"
-QUESTION_SOURCE_QUIZ = "quiz"
-QUESTION_SOURCE_CHOICES = [
-    (QUESTION_SOURCE_QUESTION, "Question"),
-    (QUESTION_SOURCE_QUIZ, "Quiz"),
-]
-
 QUIZ_RELATIONSHIP_LIST = [
     "suivant",
     # "précédent",
     "jumeau",
     "similaire",
 ]
-
-DEFAULT_DAILY_STAT_HOUR_SPLIT = {
-    str(h): {
-        "question_answer_count": 0,
-        "question_answer_from_quiz_count": 0,
-        "quiz_answer_count": 0,
-        "question_feedback_count": 0,
-        "question_feedback_from_quiz_count": 0,
-        "quiz_feedback_count": 0,
-    }
-    for h in range(24)  # '0' à '23'
-}
-
-AGGREGATION_FIELD_CHOICE_LIST = [
-    "question_answer_count",
-    # "quiz_answer_count",
-    "question_feedback_count",
-    # "quiz_feedback_count",
-]
-AGGREGATION_QUIZ_FIELD_CHOICE_LIST = ["quiz_answer_count", "quiz_feedback_count"]
-AGGREGATION_SCALE_CHOICE_LIST = ["day", "week", "month"]
-AGGREGATION_SINCE_CHOICE_LIST = ["total", "last_30_days", "month", "week"]
-AGGREGATION_SINCE_DATE_DEFAULT = "2020-01-01"
 
 CONTRIBUTION_TYPE_LIST = [
     "nouvelle question",
@@ -118,7 +81,3 @@ NOTION_QUESTIONS_IMPORT_SCOPE_CHOICES = [
 NOTION_QUESTIONS_IMPORT_SCOPE_LIST = [
     value for (value, label) in NOTION_QUESTIONS_IMPORT_SCOPE_CHOICES
 ]
-
-
-def daily_stat_hour_split_jsonfield_default_value():
-    return DEFAULT_DAILY_STAT_HOUR_SPLIT
