+++ backend/api/serializers.py
@@ -4,13 +4,8 @@ from api.models import (
     Category,
     Tag,
     Question,
-    QuestionAggStat,
-    QuestionAnswerEvent,
-    QuestionFeedbackEvent,
     Quiz,
     QuizQuestion,
-    QuizAnswerEvent,
-    QuizFeedbackEvent,
     Contribution,
     Glossary,
 )
@@ -114,45 +109,6 @@ class QuestionFullObjectSerializer(serializers.ModelSerializer):
 
 
 """
-QUESTION AGG STATS
-"""
-
-
-class QuestionAggStatSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = QuestionAggStat
-        fields = [
-            "question_id",
-            "answer_count",
-            "answer_success_count",
-            "like_count",
-            "dislike_count",
-        ]
-
-
-"""
-QUESTION ANSWER EVENT
-"""
-
-
-class QuestionAnswerEventSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = QuestionAnswerEvent
-        fields = ["question_id", "choice", "source", "created"]
-
-
-"""
-QUESTION FEEDBACK EVENT
-"""
-
-
-class QuestionFeedbackEventSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = QuestionFeedbackEvent
-        fields = ["question_id", "choice", "source", "created"]
-
-
-"""
 QUIZ QUESTION
 """
 
@@ -209,28 +165,6 @@ class QuizFullSerializer(serializers.ModelSerializer):
     class Meta:
         model = Quiz
         fields = QUIZ_FIELDS
-
-
-"""
-QUIZ ANSWER EVENT
-"""
-
-
-class QuizAnswerEventSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = QuizAnswerEvent
-        fields = ["quiz_id", "question_count", "answer_success_count", "created"]
-
-
-"""
-QUIZ FEEDBACK EVENT
-"""
-
-
-class QuizFeedbackEventSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = QuizFeedbackEvent
-        fields = ["quiz_id", "choice", "created"]
 
 
 """
