+++ backend/api/templates/admin/index_with_export.html
@@ -25,7 +25,7 @@
   <hr style="margin-top:20px;margin-bottom:20px" />
 
   <h1>Stats d'usage ?</h1>
-  <p>Voir la page <a href="{% url 'admin:api_dailystat_changelist' %}">Daily Stats</a>.</p>
+  <p>Voir la page <a href="{% url 'admin:stats_dailystat_changelist' %}">Daily Stats</a>.</p>
 
   <hr style="margin-top:20px;margin-bottom:20px" />
