+++ backend/app/urls.py
@@ -1,13 +1,16 @@
+from django.contrib import admin
 from django.urls import path, include
 
 from app import views
-from api.admin import admin_site
+
+# from api.admin import admin_site
 
 
 urlpatterns = [
     path("", views.app_home),
-    path("admin/", admin_site.urls),
+    path("admin/", admin.site.urls),  # admin_site.urls
     path("api/", include("api.urls")),
+    path("stats/", include("stats.urls")),
     path("actions/aggregate-stats", views.action_aggregate_stats),
     path(
         "actions/import-questions-from-notion",
