+++ backend/app/views.py
@@ -13,7 +13,7 @@ def app_home(request):
     return HttpResponse(
         """
         <p>Welcome to the 'Know Your Planet' Backend.</p>
-        <p>The api is available at /api</p>
+        <p>Available endpoints: /api & /stats</p>
     """
     )
