+++ backend/stats/management/commands/reset_question_stats.py
