+++ frontend/.env
@@ -1,3 +1,4 @@
 NODE_ENV=production
 VUE_APP_DOMAIN_URL=https://quizanthropocene.fr
 VUE_APP_API_ENDPOINT=https://quiz-anthropocene.osc-fr1.scalingo.io/api
+VUE_APP_STATS_ENDPOINT=https://quiz-anthropocene.osc-fr1.scalingo.io/stats
