+++ frontend/.env.example
@@ -1,3 +1,4 @@
 NODE_ENV=development  # production in prod
 VUE_APP_DOMAIN_URL=http://localhost:8000
 VUE_APP_API_ENDPOINT=http://localhost:8000/api
+VUE_APP_STATS_ENDPOINT=http://localhost:8000/stats
